requirements.txt
flask>=3.0.0
langchain>=0.2.11
langchain-community>=0.2.10
langchain-openai>=0.1.20
langgraph>=0.2.34
chromadb>=0.5.5
rank-bm25>=0.2.2
tiktoken>=0.7.0
pydantic>=2.7.0
python-dotenv>=1.0.1
ragas>=0.1.9
datasets>=2.19.0
openai>=1.40.0

